+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.0.2
+
+- Use deterministic deployments in migrations.
+- Update dependency version `@animoca/ethereum-contracts@0.0.5`.
+
 ## 0.0.1
 
 - Initial release.